+++ service/verify/v2/service/entities/api_op_page.go
@@ -1,0 +1,174 @@
+// Package entities contains auto-generated files. DO NOT MODIFY
+package entities
+
+import (
+	"context"
+	"net/http"
+	"net/url"
+	"strconv"
+	"time"
+
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/utils"
+)
+
+// EntitiesPageOptions defines the query options for the api operation
+type EntitiesPageOptions struct {
+	PageSize  *int
+	Page      *int
+	PageToken *string
+}
+
+type PageEntityResponse struct {
+	AccountSid  string     `json:"account_sid"`
+	DateCreated time.Time  `json:"date_created"`
+	DateUpdated *time.Time `json:"date_updated,omitempty"`
+	Identity    string     `json:"identity"`
+	ServiceSid  string     `json:"service_sid"`
+	Sid         string     `json:"sid"`
+	URL         string     `json:"url"`
+}
+
+type PageMetaResponse struct {
+	FirstPageURL    string  `json:"first_page_url"`
+	Key             string  `json:"key"`
+	NextPageURL     *string `json:"next_page_url,omitempty"`
+	Page            int     `json:"page"`
+	PageSize        int     `json:"page_size"`
+	PreviousPageURL *string `json:"previous_page_url,omitempty"`
+	URL             string  `json:"url"`
+}
+
+// EntitiesPageResponse defines the response fields for the entities page
+type EntitiesPageResponse struct {
+	Entities []PageEntityResponse `json:"entities"`
+	Meta     PageMetaResponse     `json:"meta"`
+}
+
+// Page retrieves a page of entities
+// See https://www.twilio.com/docs/verify/api/entity#read-multiple-entity-resources for more details
+// Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+func (c Client) Page(options *EntitiesPageOptions) (*EntitiesPageResponse, error) {
+	return c.PageWithContext(context.Background(), options)
+}
+
+// PageWithContext retrieves a page of entities
+// See https://www.twilio.com/docs/verify/api/entity#read-multiple-entity-resources for more details
+func (c Client) PageWithContext(context context.Context, options *EntitiesPageOptions) (*EntitiesPageResponse, error) {
+	op := client.Operation{
+		Method: http.MethodGet,
+		URI:    "/Services/{serviceSid}/Entities",
+		PathParams: map[string]string{
+			"serviceSid": c.serviceSid,
+		},
+		QueryParams: utils.StructToURLValues(options),
+	}
+
+	response := &EntitiesPageResponse{}
+	if err := c.client.Send(context, op, nil, response); err != nil {
+		return nil, err
+	}
+	return response, nil
+}
+
+// EntitiesPaginator defines the fields for makings paginated api calls
+// Entities is an array of entities that have been returned from all of the page calls
+type EntitiesPaginator struct {
+	options  *EntitiesPageOptions
+	Page     *EntitiesPage
+	Entities []PageEntityResponse
+}
+
+// NewEntitiesPaginator creates a new instance of the paginator for Page.
+func (c *Client) NewEntitiesPaginator() *EntitiesPaginator {
+	return c.NewEntitiesPaginatorWithOptions(nil)
+}
+
+// NewEntitiesPaginatorWithOptions creates a new instance of the paginator for Page with options.
+func (c *Client) NewEntitiesPaginatorWithOptions(options *EntitiesPageOptions) *EntitiesPaginator {
+	return &EntitiesPaginator{
+		options: options,
+		Page: &EntitiesPage{
+			CurrentPage: nil,
+			Error:       nil,
+			client:      c,
+		},
+		Entities: make([]PageEntityResponse, 0),
+	}
+}
+
+// EntitiesPage defines the fields for the page
+// The CurrentPage and Error fields can be used to access the PageEntityResponse or error that is returned from the api call(s)
+type EntitiesPage struct {
+	client *Client
+
+	CurrentPage *EntitiesPageResponse
+	Error       error
+}
+
+// CurrentPage retrieves the results for the current page
+func (p *EntitiesPaginator) CurrentPage() *EntitiesPageResponse {
+	return p.Page.CurrentPage
+}
+
+// Error retrieves the error returned from the page
+func (p *EntitiesPaginator) Error() error {
+	return p.Page.Error
+}
+
+// Next retrieves the next page of results.
+// Next will return false when either an error occurs or there are no more pages to iterate
+// Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+func (p *EntitiesPaginator) Next() bool {
+	return p.NextWithContext(context.Background())
+}
+
+// NextWithContext retrieves the next page of results.
+// NextWithContext will return false when either an error occurs or there are no more pages to iterate
+func (p *EntitiesPaginator) NextWithContext(context context.Context) bool {
+	options := p.options
+
+	if options == nil {
+		options = &EntitiesPageOptions{}
+	}
+
+	if p.CurrentPage() != nil {
+		nextPage := p.CurrentPage().Meta.NextPageURL
+
+		if nextPage == nil {
+			return false
+		}
+
+		parsedURL, err := url.Parse(*nextPage)
+		if err != nil {
+			p.Page.Error = err
+			return false
+		}
+
+		options.PageToken = utils.String(parsedURL.Query().Get("PageToken"))
+
+		page, pageErr := strconv.Atoi(parsedURL.Query().Get("Page"))
+		if pageErr != nil {
+			p.Page.Error = pageErr
+			return false
+		}
+		options.Page = utils.Int(page)
+
+		pageSize, pageSizeErr := strconv.Atoi(parsedURL.Query().Get("PageSize"))
+		if pageSizeErr != nil {
+			p.Page.Error = pageSizeErr
+			return false
+		}
+		options.PageSize = utils.Int(pageSize)
+	}
+
+	resp, err := p.Page.client.PageWithContext(context, options)
+	p.Page.CurrentPage = resp
+	p.Page.Error = err
+
+	if p.Page.Error == nil {
+		p.Entities = append(p.Entities, resp.Entities...)
+	}
+
+	return p.Page.Error == nil
+}